+++ node/src/net/mod.rs
@@ -9,6 +9,7 @@ pub mod manager;
 pub mod peer;
 pub mod permissions;
 pub mod proxy;
+pub mod stats;
 pub mod tor;
 pub mod sync;
 pub mod v2transport;